+++ src/ContaoManager/Plugin.php
@@ -12,6 +12,7 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Contao\ManagerPlugin\Config\ConfigPluginInterface;
 use Contao\ManagerPlugin\Config\ContainerBuilder;
 use Contao\ManagerPlugin\Config\ExtensionPluginInterface;
 use Contao\ManagerPlugin\Routing\RoutingPluginInterface;
@@ -19,11 +20,12 @@ use HeimrichHannot\EncoreBundle\HeimrichHannotContaoEncoreBundle;
 use HeimrichHannot\IsotopeBundle\HeimrichHannotContaoIsotopeBundle;
 use HeimrichHannot\SlickBundle\HeimrichHannotContaoSlickBundle;
 use HeimrichHannot\UtilsBundle\Container\ContainerUtil;
+use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\Config\Loader\LoaderResolverInterface;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Component\Routing\RouteCollection;
 
-class Plugin implements BundlePluginInterface, ExtensionPluginInterface, RoutingPluginInterface
+class Plugin implements BundlePluginInterface, ExtensionPluginInterface, RoutingPluginInterface, ConfigPluginInterface
 {
     /**
      * {@inheritdoc}
@@ -49,23 +51,52 @@ class Plugin implements BundlePluginInterface, ExtensionPluginInterface, Routing
      */
     public function getExtensionConfig($extensionName, array $extensionConfigs, ContainerBuilder $container)
     {
-        $extensionConfigs = ContainerUtil::mergeConfigFile('huh_list', $extensionName, $extensionConfigs, __DIR__.'/../Resources/config/config_list.yml');
+        $extensionConfigs = ContainerUtil::mergeConfigFile(
+            'huh_list',
+            $extensionName,
+            $extensionConfigs,
+            __DIR__.'/../Resources/config/config_list.yml'
+        );
 
-        $extensionConfigs = ContainerUtil::mergeConfigFile('huh_reader', $extensionName, $extensionConfigs, __DIR__.'/../Resources/config/config_reader.yml');
+        $extensionConfigs = ContainerUtil::mergeConfigFile(
+            'huh_reader',
+            $extensionName,
+            $extensionConfigs,
+            __DIR__.'/../Resources/config/config_reader.yml'
+        );
 
-        return ContainerUtil::mergeConfigFile('huh_encore', $extensionName, $extensionConfigs, __DIR__.'/../Resources/config/config_encore.yml');
+        return ContainerUtil::mergeConfigFile(
+            'huh_encore',
+            $extensionName,
+            $extensionConfigs,
+            __DIR__.'/../Resources/config/config_encore.yml'
+        );
     }
 
     /**
      * Returns a collection of routes for this bundle.
      *
      * @param LoaderResolverInterface $resolver
-     * @param KernelInterface         $kernel
+     * @param KernelInterface $kernel
      *
      * @return RouteCollection|null
+     * @throws \Exception
      */
     public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel)
     {
         return $resolver->resolve(__DIR__.'/../Resources/config/routing.yml')->load(__DIR__.'/../Resources/config/routing.yml');
+    }
+
+    /**
+     * Allows a plugin to load container configuration.
+     *
+     * @param LoaderInterface $loader
+     * @param array $managerConfig
+     * @throws \Exception
+     */
+    public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
+    {
+        $loader->load('@HeimrichHannotContaoIsotopeBundle/Resources/config/services.yml');
+        $loader->load('@HeimrichHannotContaoIsotopeBundle/Resources/config/listeners.yml');
     }
 }